requirements.txt
Django==4.2.13
django-colorfield==0.11.0
django-fontawesome-5==1.0.18
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
pillow==10.3.0
tablib==3.5.0
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==5.0.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.2
cssselect2==0.7.0
daphne==4.1.0
defusedxml==0.7.1
diff-match-patch==20230430
django-bleach==3.1.0
django-ckeditor==6.7.0
django-comments-dab==2.8.0
django-import-export==3.2.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-tinymce==3.6.1
docx==0.2.4
et-xmlfile==1.1.0
flower==2.0.1
gunicorn==21.2.0
html5lib==1.1
htmldocx==0.0.6
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.5
lxml==5.1.0
MarkupPy==1.14
msgpack==1.0.8
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyOpenSSL==24.0.0
pypandoc==1.12
pypdf==4.0.1
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.2
reportlab==4.0.6
requests==2.31.0
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tornado==6.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.2

